requirements.txt
python-binance
Ta-Lib
numpy
websocket_client
